requirement.txt
python-dotenv
requests
psycopg2-binary
openai
transformers
torch
anthropic
accelerate
bitsandbytes
pandas